first cut: server works, but has issues
In this version of the new extractor magic, the server works.  However,
I made a few compromises that are at least unfortunate and at worst
untenable.

The distateful-but-not-necessarily-dealbreaker issue is that in order to
implement ApiHandler for a function type that itself contains a generic
parameter (Query<Q> as an argument to the function), I need to
parametrize the implementation (impl<..., Q> ApiHandler<...> ..., where
the "Q" shows up in the function's type signature).  However, Rust
doesn't allow you to have a type parameter that isn't part of the "self"
type.  I think this basically _is_ part of the self type (because "self"
is a function that consumes the Query); however, the self type doesn't
own or refer to an instance of Q or &Q or anything.  (Is that the
problem?) I dealt with this by wrapping such functions in a
FuncQueryWrapper and using a PhantomData<Q>.  This _does_ work, which
makes me feel I understood the above issue (namely: a function taking a
Query<Q> does not "count" has constraining the trait bound Q).

The bigger issue is a compromise that I made much earlier.  This
solution is premised on being able to implement ApiHandler differently
for various different function types (i.e., based on their type
signature).  However, Rust does not consider trait bounds in determining
if there are conflicting implementations; as a result, the different
implementations of ApiHandler all appear to conflict.  I worked around
this by adding a type parameter to ApiHandler that's effectively a
marker type -- each different implementation uses a different type.  The
problem with this is that the marker type leaks into the type of the
handler.  The end result is that you cannot have a trait object of type
ApiHandler -- it has to be a _specific_ ApiHandler<T>.  That's not a
non-starter for extractors per se, but it _is_ a nonstarter for being
able to shove these into a router data structure and be able to invoke
them uniformly, which was part of the point here.