remove ApiBackend and Simulator
As the implementation is fleshed out, the vast majority of code in the
Simulator is just managing data structures representing the state of the
rack.  This code is all common to both a simulated backend and a real
one.  This has been moved into a new OxideRack interface, which is an
HTTP-agnostic representation of the interface to the system.  (A
subsequent step will likely be to move the data representation parts
into a separate data storage interface.)

There is no real simulation per se yet (nor was there before this
change).  One of the next steps will be to introduce that by simulating
entire compute servers.  This appears to make more sense than having a
Backend interface with a simulated vs. real implementation.  With the
previous Backend interface, it was somewhat ambiguous what code ought to
be shared between the simulated vs. real backend.  With a simulated
compute server, the only question is whether the code ought to belong in
this program or the compute server, which seems a bit clearer.