Adds exponential backoff to the retry loop of the sled agent
- Adds module constants for the maximum retry and the backoff factor.
Constants which represent time have been defined as
`std::time::Duration` objects to avoid confusion about units and most
conversions.
- Adds a small type `RetryBackoff`, which implements the backoff
strategy. This includes the initial/maximal delays, the randomization
factor, the backoff factor, and the maximum number of retries. The type
is intended to be used as an interator, calling `next()` to retrieve the
next delay.
- Adds tests of the above type.