Address code review comments
- Moves util.rs -> backoff.rs
- Removes custom implementation of backoff, and uses the third-party
backoff crate instead, rexposing the most useful symbols from that
crate.
- Adds a "retry policy" function for connecting to internal Oxide
services. This policy is an infinite retry loop, with a long maximum
interval of 60 minutes. This is the only policy at this point, but
others may be added to express how user-facing services will be retried.
The goal here is to provide (1) some consistent, simple defaults for
various use-cases, and (2) to hide some of the sharp edges of the
`backoff` crate, which uses `f64`s in a lot of places and with which
it's pretty easy to get into trouble (e.g., _decreasing_ retry
intervals).