Adds exponential backoff to the retry loop of the sled agent (#33)
* Adds exponential backoff to the retry loop of the sled agent

- Adds module constants for the maximum retry and the backoff factor.
Constants which represent time have been defined as
`std::time::Duration` objects to avoid confusion about units and most
conversions.
- Adds a small type `RetryBackoff`, which implements the backoff
strategy. This includes the initial/maximal delays, the randomization
factor, the backoff factor, and the maximum number of retries. The type
is intended to be used as an interator, calling `next()` to retrieve the
next delay.
- Adds tests of the above type.

* Address code review comments

- Moves util.rs -> backoff.rs
- Removes custom implementation of backoff, and uses the third-party
backoff crate instead, rexposing the most useful symbols from that
crate.
- Adds a "retry policy" function for connecting to internal Oxide
services. This policy is an infinite retry loop, with a long maximum
interval of 60 minutes. This is the only policy at this point, but
others may be added to express how user-facing services will be retried.
The goal here is to provide (1) some consistent, simple defaults for
various use-cases, and (2) to hide some of the sharp edges of the
`backoff` crate, which uses `f64`s in a lot of places and with which
it's pretty easy to get into trouble (e.g., _decreasing_ retry
intervals).

* Addresses minor PR issues

- Reduces initial retry duration
- Clippy lint fix
- Reorder dependencies in Cargo.toml