Starting refactor to a more statically-typed interface
- Removes integration tests for now
- Adds traits module

This starts work towards a more statically-typed set of traits, by
adding an associated type to the `Metric` trait that represents the of
data a metric is supposed to generated. It also seems like we'll remove
the target and metric types from the Collector's interface -- the client
almost certainly needs to keep these themselves, to know which ones to
modify as the targets change. The Collector is really just the interface
for collecting the measurements or samples from a timeseries, not for
storing the targets/metrics themselves. Helper types can probably be
provided for doing that, e.g., beefing up the Counter type to store a
counter for a _set_ of metrics.