Improves representation of bins in histogram
Previously, we used `RangeInclusive` to store bin edges. This is wrong,
because the right edge of a bin overlaps the left edge of the next. This
commit adds a dedicated type for storing bin ranges, which allows three
different variants: unbounded below and exclusive above; inclusive below
and exclusive above; and inclusive below and unbounded above. This
enables handling the entire support of both i64 and f64 and uniquely
assign any sample (except for NaNs) to a bin. We can also support +/-
infinity, although it's still not clear if that's the right approach.

This update means that the original constructor for Histogram accepts
has been changed to accept the _left_ bin edges, which are always
inclusive. It also allows users to bring their own bins, by specifying
normal Rust range types and converting them. (E.g., `(0..10).into()`.)

The documentation has also been updated to more explicitly describe how
the bin edges and bins are handled and when errors are returned.
Examples and tests have also been added for the new behavior.