Much simpler crate structure
This commit unifies the three macro implementation crates into one. This
overwrought structure was really there because the `MeasurementType` was
defined in the macro crate, for a few historical and mostly incorrect
reasons. By moving this into the `oximeter` crate we need to operate on
an Ident in the macro crate rather than the type, but this means we can
put all the macros together.