Integrates existing oximeter repository into omicron (#79)
* Initial commit of metric production library

- Adds procedural macros for decorating small Rust structs as a Target
  or a Metric. These structs are essentially the schema for a
  timeseries. The derived traits provide methods for introspecting the
  structs, such as listing their name and the names/type/values of their
  fields.
- Adds a `Collector` type, which aggregates metrics in client code.
  Users register their targets and metrics with the collector, along
  with a `Producer` type. The `Producer` trait allows the `Collector` to
  retrieve the current values of its metrics, whenever its `collect`
  method is called. This provides some flexibility: users can define a
  type that has state or other logic, and the implement the `Producer`
  trait for it, which allows them some control over how the metrics are
  actually collected.
- Metrics can be registered as a "collection" -- just a set of metrics
  that are coupled in some way. The `Producer` trait has a method for
  setting up a pulling the metrics from a collection, and one for
  pulling all the metrics in sequence. Again, this affords some
  flexibility to the client to decide how the metrics are actually
  generated.
- Adds two concrete helper types that implement `Producer`, likely to be
  the most common. One is an atomic `Counter`, which just manages a
  single i64 counter and allows users to bump that whenever they want.
  The other is a `Distribution`, which allows users to add samples
  whenever want.

* Adds runtime type-checking of the measurements during each collection

* Adds comments

This commit adds a bunch of comments, mostly to show how the thing is
currently used, and what the thinking is behind the various types.

* Starting refactor to a more statically-typed interface

- Removes integration tests for now
- Adds traits module

This starts work towards a more statically-typed set of traits, by
adding an associated type to the `Metric` trait that represents the of
data a metric is supposed to generated. It also seems like we'll remove
the target and metric types from the Collector's interface -- the client
almost certainly needs to keep these themselves, to know which ones to
modify as the targets change. The Collector is really just the interface
for collecting the measurements or samples from a timeseries, not for
storing the targets/metrics themselves. Helper types can probably be
provided for doing that, e.g., beefing up the Counter type to store a
counter for a _set_ of metrics.

* Adds associated type to the Metric trait

- Adds an associated type, Measurement, to the Metric trait. This will
  be used to constrain the types of a producer later.
- Renames MetricType -> MeasurementType
- Adds the Cumulative types to the MeasurementType and impls DataPoint
  for them as well
- Simplifies the `#[oximeter::metric]` macro to accept _only_ the type
  that will be used as the associated `Metric::Measurement` type.

* Adds concrete types used to represent a sample from a timeseries

- Adds the structs `types::{Target, Metric, Sample}` to store a
  timestamped measurement from a single target and metric. The choice of
  "duplicated" types currently seems better than either a generic sample
  type, or one operating only on boxed trait objects. The latter two
  have drawbacks around type-safety, performance, and generality, and
  this approach makes the most sense at this point.
- Simplifies the producer trait, just a single method to yield an
  iterator over `Sample`s. They can be of any measurement type and from
  any target or metric (one key benefit of the approach described
  above).
- Adds back some integration tests using these new types.

* Moves integration tests into units

- Adds a workaround that allows the macros from this crate to be used
  inside it, by using `extern crate self as oximeter`. See the commit
  for related notes and links.
- Moves the integration tests into the library as unit tests, using the
  above fact. Tests for the target/metric/producer traits are in
  `src/traits.rs` and the structs are in `src/types.rs`.

* Adds docs and readme, and some small cleanup

* It's a histogram, not a distribution.

Also adds `::` to types emitted from the proc macros.

* Adds comparison operations and a producer error

Comparisons will be helpful when storing measurements in a reasonable
order for delivery to the client.

* Feeding the clippy monster

* Adds oximeter macro implementation crate to the default workspace members, so that it is tested automatically

* Improves representation of bins in histogram

Previously, we used `RangeInclusive` to store bin edges. This is wrong,
because the right edge of a bin overlaps the left edge of the next. This
commit adds a dedicated type for storing bin ranges, which allows three
different variants: unbounded below and exclusive above; inclusive below
and exclusive above; and inclusive below and unbounded above. This
enables handling the entire support of both i64 and f64 and uniquely
assign any sample (except for NaNs) to a bin. We can also support +/-
infinity, although it's still not clear if that's the right approach.

This update means that the original constructor for Histogram accepts
has been changed to accept the _left_ bin edges, which are always
inclusive. It also allows users to bring their own bins, by specifying
normal Rust range types and converting them. (E.g., `(0..10).into()`.)

The documentation has also been updated to more explicitly describe how
the bin edges and bins are handled and when errors are returned.
Examples and tests have also been added for the new behavior.

* Feeding the clippy monster

* Much simpler crate structure

This commit unifies the three macro implementation crates into one. This
overwrought structure was really there because the `MeasurementType` was
defined in the macro crate, for a few historical and mostly incorrect
reasons. By moving this into the `oximeter` crate we need to operate on
an Ident in the macro crate rather than the type, but this means we can
put all the macros together.