Adds an actual oximeter metric collection server
This commit adds the "collection side" of the metrics collection system.
The `oximeter::Oximeter` struct is the main server exposing the
interface of the internal agent that does the scraping, `OximeterAgent`,
which is private.

Adds the POST endpoint in Nexus for handling registration requests of a
new oximeter instance, similar to the method the sled agent itself uses.