Adds packaging for oximeter
Adds the main `oximeter` binary to the list of those packaged by
`omicron-package`, along with an SMF manifest for it. Also updates the
`config.toml` file that's bundled with `nexus`, which seems to have been
stale (missing fields like ID, etc). This all unpacks and runs as
expected under /opt/oxide.