Moves the actual metric data value into the derived struct itself.
Previously, the `oximeter::metric` attribute macro required an
identifier that specified the metric data to be produced. However, the
consumer of the code was required to keep the target and metric structs,
_and_ the actual measured metric values in some other type as well. This
meant there were at least three types floating around that really should
be tied together.

This commit changes the `oximeter::Metric` trait, and provides a
derive-macro for it, instead of the attribute macro. The struct that
derives that must have at least one field, named `value`, which must be
one of the supported metric data types. The others are used to generate
the field-name/field-value lists as before, and a new method is used to
derive an `oximeter::Measurement` from the current `value` field. This
also simplifies the `oximeter:types::Sample` type, which now only needs
the target and metric, and pulls the measurement from the metric as
well.