Fixes bug in oximeter::Metric macro implementation
There was a bug expanding the key of a metric, which only manifested
when there are zero fields. That was previously unsupported, but now is,
hence the arrival of the bug. I've also added some tests of the compiler
error messages, which should help catch regressions or other similar
such issues.