Adds schema validation and update to the ClickHouse client
Adds tools to validate schema of each sample as it is received and
inserted. The client caches the target/metric schema when it is created.
As samples come in, they are validated against those schema. Anything
_violating_ them is skipped. If the schema don't exist, they are
inserted into the local cache and the database.

Also includes ClickHouse tests behind a platform/feature-flag, and adds
a shell script to install the server on CI.