Simplifies some of the serialization with ClickHouse
Adds a set of types for representing data as ClickHouse expects it, and
methods for converting between the preferred representation in Rust and
ClickHouse. As an example, the `Field` type is used to store the name
and type of a single field in a target/metric schema. We store a vector
of these in the schema itself. ClickHouse needs these to be an array of
structs, rather than a struct of arrays, so the `DbFieldList` type is
used to represent the latter. The `Db*` types in general are used to
communicate with ClickHouse while the types without those prefixes are
used in Rust.

Also adds the `Schematized` trait, to reduce some code duplication when
handling the `{Target,Metric}Schema` objects.