Adds schema validation and update to the ClickHouse client (#110)
* Adds schema validation and update to the ClickHouse client

Adds tools to validate schema of each sample as it is received and
inserted. The client caches the target/metric schema when it is created.
As samples come in, they are validated against those schema. Anything
_violating_ them is skipped. If the schema don't exist, they are
inserted into the local cache and the database.

Also includes ClickHouse tests behind a platform/feature-flag, and adds
a shell script to install the server on CI.

* Simplifies some of the serialization with ClickHouse

Adds a set of types for representing data as ClickHouse expects it, and
methods for converting between the preferred representation in Rust and
ClickHouse. As an example, the `Field` type is used to store the name
and type of a single field in a target/metric schema. We store a vector
of these in the schema itself. ClickHouse needs these to be an array of
structs, rather than a struct of arrays, so the `DbFieldList` type is
used to represent the latter. The `Db*` types in general are used to
communicate with ClickHouse while the types without those prefixes are
used in Rust.

Also adds the `Schematized` trait, to reduce some code duplication when
handling the `{Target,Metric}Schema` objects.

* Addressing PR comments