Adding tools to manage ClickHouse during development
- Updates tool for downloading ClickHouse, now using prebuilt binaries
from S3.
- Adds the `omicron_common::dev::clickhouse::ClickHouseInstance` object.
This is modeled after the existing `CockroachInstance` object, and is
used to start and manage a ClickHouse server sub-process. It currently
is quite constrained, only allowing users to specify the HTTP port the
server listens on. This is for simplicity, and further options may be
needed in the future.
- Updates the Oximeter client tests. These now spawn their own
`ClickHouseInstance` in each test, so they can be run in parallel. This
also removes the feature flag used to hide the tests on non-linux
platforms.
- Updates the GitHub workflow to use the new tool to download ClickHouse
- Adds the `ch-run` subcommand to the `omicron-dev` tool, and updates
README.adoc with details about how to run