Store metric endpoint information in CockroachDB (#127)
* Stores metric producer/collector information CockroachDB

- Adds tables to the omicron database to store oximeter collectors,
metric producers, and the assignment of producers to collectors
- Adds Rust types and conversion methods for reading/writing records
from these new tables
- Updates the existing Nexus endpoint for registering an oximeter
collector instance into the appropriate table
- Updates the existing Nexus endpoint for registering a metric producer
to: insert the producer information into the appropriate table; select
an oximeter instance (currently from memory) for assignment; insert the
assignment into the table; complete the assignment as before.

* Adds a few integration tests for Oximeter

- Adds an Oximeter collector server and producer server to the Nexus
common integration test context.
- Verifies that records for Oximeter, the producer, and the assignment
between them are accurately stored in the control plane database.

* Adds better integration tests and support for automatic port assignment

During integration tests, we want to run multiple copies of various
servers side-by-side. This means we can't hard-code ports. This commit
adds support for automatically assigning ports to ClickHouse and the
`oximeter::ProducerServer` type.

For the latter, the `ProducerEndpoint` type that's used to set up the
`ProducerServer` can optionally have a socket address with port 0. In
this case, that value is _overwritten_ with the Dropshot HTTP server's
local port, once it's been bound and listening. A log message is printed
in this case.

For the ClickHouse server, this requires some annoying workarounds.
ClickHouse supports listening on port 0, but it doesn't give us a way to
discover the actual port the OS assigned it. At this point, we have to
ask the OS. That's currently done with the `lsof` on most platforms, and
`pfiles` on illumos, but in both cases it implies some shitty parsing of
the text output. This is all done (and tested) in
`omicron_common::dev::clickhouse`. Note that all tests of the parsing
are run on all platforms -- the expected output is included in the
tests, not collected at runtime from the system.

This commit also moves some types around in `oximeter`. The `collect`
module now has _just_ the `Collector` type, and the producer server and
its associated types are in `oximeter::producer_server`. The
`ProducerServer` itself is re-exported at the crate root.

* Addressing PR comments