Parses clickhouse log file for discovering port number
Removes port-discovery methods parsing `lsof`/`pfiles` output, and
parses ClickHouse log file for the HTTP port it listens on. This is only
done if the program is run with a port of 0, so the OS assigns any
available port. This relies on upstream changes in ClickHouse -- and the
script to download ClickHouse has updated MD5s to reflected that.