Fixes race in detection of ClickHouse port from log file
The original implementation of this code searched the ClickHouse log
file for the port number, but only until EOF. There's a race here, as
the ClickHouse subprocess may not write the port number by the time the
reading code starts parsing the log -- the reader hits EOF before the
port is written, and bails.

This commit adds a loop for reading the log file that ignores EOF,
sleeping for a time. This is run with a timeout to avoid an infinite
loop. Regression tests have also been added.