Fixes race in detection of ClickHouse port from log file (#139)
* Fixes race in detection of ClickHouse port from log file

The original implementation of this code searched the ClickHouse log
file for the port number, but only until EOF. There's a race here, as
the ClickHouse subprocess may not write the port number by the time the
reading code starts parsing the log -- the reader hits EOF before the
port is written, and bails.

This commit adds a loop for reading the log file that ignores EOF,
sleeping for a time. This is run with a timeout to avoid an infinite
loop. Regression tests have also been added.

* Avoids explicit sleeping in ClickHouse log-parsing tests

Uses `tokio`'s utilities for pausing and advancing time explicitly to
run the tests for parsing ClickHouse's log file for a port number. These
can be used to avoid long sleeps, as we can directly advance time
"beyond" those points, by an amount appropriate for the test

* Addressing PR comments and feature-resolution fixes

* Adds guard for tokio::select! branch to avoid double-polling task

* Use interior mutability to guarantee the temp logfile outlives both the
reading and writing tasks.