When propolis says 'destroyed', the Sled Agent thinks 'stopped'. (#136)
When Propolis self-reports that local state has been destroyed,
the Sled Agent should report to Nexus that the state has been
stopped.

This is a matter of semantics, but the distinction is important:
when an instance is stopped within Propolis, it will never again
run on that specific server. However, the instance (and associated
virtual resources) may be re-allocated elsewhere - so they are
"stopped" from a client perspective.