Use OS-assigned port in oximeter DB client tests
- Removes hard-coded port numbers in tests of the oximeter ClickHouse
client, opting for OS-assigned ports to really avoid collisions.
- Simplifies internals of ClickHouse runner. Previously, we checked for
the a "status" file, which the ClickHouse binary spits out with
information like the PID. But this is _before_ the server(s) start,
which means it was possible that the server's socket is not
bound/listening by the time we actually query it, _in the case that we
start it with a known port_. If we use an OS-assigned port, we wait for
the log-file to tell us the actual bound port, which means the server
has to actually be listening at that point. This change now
unconditionally parses the log file for the port, meaning that we're
always guaranteed that the server is ready for connections when we
return from `ClickHouseInstance::new`. This avoids _yet another_ race in
the startup/management of the ClickHouse subprocess.