Removes implicit ordering in materialized targets and metrics (#145)
The `oximeter::types::{Target,Metric}` structs previously stored a
B-tree mapping field names to field values. This collection sorts its
keys, so there's possibly a different ordering between the definition of
the fields (which is how they're stored in the DB) and the fields as
stored in those structs. This came up while implementing
filtering/querying on the DB, in which case, rebuilding a `Target` or
`Metric` struct from the values in the timeseries key no longer works,
since it requires the ordering be preserved.

This commit removes any B-trees used to store name/value pairs of a
field, and introduces the `oximeter::types::Field` struct which combines
the name and value into one. Linear sequences of these (vectors, slices,
etc) are used in place of maps.