Unifies targets and metrics early in the sampling process
Metrics and targets are used to distinguish the features we're measuring
from the resources we're monitoring. But the distinction between targets
and metrics gets less important when querying, which I discovered during
the process of actually implementing querying and filtering. For
example, when searching for a timeseries, we want to filter by
field=value pairs, but we don't particularly care whether those fields
apply to the target or metric.

This commit collapses the distinction between targets and metrics during
the process of generating a `Sample`, transforming them into a name and
an ordered sequence of field name/value pairs. The representation in the
database has also been simplified a lot, as there are no longer tables
for each type of field, for _both_ targets and metrics, only a single
field table for each type. The tables for target and metric schema have
also been reduced to just one table for timeseries schema. This really
reduces the code size and complexity -- there was a lot of dispatching
on targets vs. metrics, often in confusing or obtuse ways.

Another important change is introduced here. The timeseries key was
previously constructed as `target_name:field1:field2:...:metric_name`.
The target and metric names have been removed from the key. A timeseries
is identified by its name, which is `target_name:metric_name`, and its
key which is `field1:field2:...`. This should also help lookup in the
database, as the keys and names are orthogonal. Since the metric name
was previously at the _end_ of the timeseries key, this really reduces
the usefulness of the key as a sorting criterion, especially when we're
also sorting by the target/metric name. Keeping them separate reduces
duplicate information in the key, and improves sorting, compression, and
lookup by timeseries key.