Improves telemetry database schema for querying (#147)
This commit changes the sorting key for many of the telemetry database
tables, most notably those storing field names/values. Querying will be
implemented by writing clauses like `WHERE field_name = 'foo' and
field_value = 'bar'`. However, the `field_value` column was not
previously in the sorting key, which forces ClickHouse to scan the
entire table for matching records. The new key includes the field value,
which enables the DB to skip records quickly. Example queries can reduce
the number of rows scanned by many orders of magnitude.

Note that while the timeseries key is still part of the index for the
measurement tables, it's also been removed from the field tables. The
key is the concatenation of the field values, which means that the
sorting implied by individual field values often conflicts with the
sorting implied by the timeseries key. At the same time, we're only
_reading_ the timeseries key of matching columns, to then select
measurements with that key, and not really using the key directly when
searching the field tables. So it's at best unhelpful, but more likely
an impediment to an efficient sort order for the way we'll be using the
tables.