Removes per-sample records in the field tables
As samples are collected, the fields for each sample are unrolled into
one record per field. So for a sample with N fields, N+1 records are
inserted into the database across all the tables, one for the
measurement and N for the samples. These are highly duplicative,
containing no new information between each record other than the
timestamp, which is already recorded in the measurement table. The
schema were originally designed this way for simplicity and because
ClickHouse provides no uniqueness constraint between records, even those
sharing the same primary or sorting key, when using its default tables.

But this comes with a cost -- there are far more fields than
measurements, so a huge amount of space is wasted storing them, as well
as the time spent actually "computing" the records. (More accurately,
cloning things like the timestamps and timeseries keys, and then
converting everything to strings for insertion.)

This commit reorganizes the database schema and insertion to insert only
a _single_ record per field _per unique timeseries_. This relies on
ClickHouse's `ReplacingMergeTree` table engine, which, rather than simply
merging data in the background, removes records with duplicated sorting
keys. This does require a slightly less efficient sorting key -- the
timeseries key has been reinstated in the sorting key, as it's required
to avoid overly-aggressive pruning of records which share the same field
values, but different timeseries keys. This inefficiency is more than
made up for by the reduced sized of the tables. The number of records
now scales with the number of unique timeseries, rather than the product
of that and the number of samples.