Adds a start time for cumulative metrics
Previously, all samples for a timeseries contained an associated
timestamp. This is incomplete, specifically for cumulative metrics.
These are by definition referenced to a starting time. This will be
important for keeping track of cumulative metrics even across things
like restarts of a monitored service, or generally something that resets
a cumulative counter.

This commit aguments `oximeter` to include a notion of a start or
reference time for the cumulative metric data types. It also adds a
sample time or period for each datum, which is either an instant (for
gauge types) or a non-zero interval (for cumulatives). The sample time
belongs properly to the `Measurement` -- it uses the cumulative data
types start time, or is defined as an instant for gauges. Several of the
types in `oximeter` have been fleshed out, clarified, and cleaned up as
well.

One changes is _how_ the `oximeter::Metric` trait is derived.
Previously, one needed the special field named `value`. To bring this
inline with the other changes to the types in `oximeter`, this field is
now called `datum`. To add flexibility, consumers may also annotated any
field they choose with the `#[datum]` attribute. These two mechanisms
are equivalent.

- Adds the `oximeter::SampleTime` type, which represents the time over
  which a measurement of a metric is sampled (an instant or an
  interval).
- Removes the unnecessary `oximeter::{Target, Metric}` structs.
- Adds the `oximeter::Datum` type, which represents a single data point
  from a metric, with no timing information.
- Changes the special field of the `Metric` struct from `value` to
  `datum`, and adds support for annotating the desired field with the
  attribute helper.
- Augments the `oximeter::Measurement` type with a contained `Datum` and
  `SampleTime`. This is now the type generated by the `Metric::measure`
  method.
- Updates ClickHouse table schema to reflect the above changes
- Updates the types modeling rows in the ClickHouse measurement tables.
  The gauge measurement tables are unchanged, but the cumulative scalar
  and histogram metrics have an additional column, `start_time`.
- Removes the now-unnecessary `oximeter::db::model::TimeseriesSample`
  type, which is now just the same as the `Measurement` type.
- Now returns the samples in a `Timeseries` as a `Vec<Measurement>`
- Includes the timeseries name in a returned `Timeseries` when querying
  the database