Fixes some oddities dealing with ClickHouse (#278)
This commit fixes two bugs dealing with ClickHouse, and adds regressions
for them.

The first is about deserialization of structs which have a `&str` field.
`serde_json` is generally fine about zero-copy, except if the string
requires escaping. In that case, it can no longer deserialize the data
into a `&str`, since it must take ownership and manipulate the returned
buffer. There are a few ways around this, but this commit opts for the
simplest -- using a owned `String` instead of a `&str`. The string is
immediately moved into a `Measurement`, which takes ownership anyway, so
this is the most straightfoward choice. See
https://github.com/serde-rs/serde/issues/1413 for more details. Note
that we also parse all string datum from the DB as `String` directly,
rather than as `&str`, again to avoid this.

The second bug is handling a ClickHouse oddity. By default, 64-bit
integer types are returned as _quoted_ strings. As in, `"1"` instead of
`1`. This is apparently to support JavaScript's JSON implementation, but
in any case it's not the behavior we expect or want. There is thankfully
a setting to turn this off, which the database client object now sets
when it makes requests returning JSON data. See
https://github.com/ClickHouse/ClickHouse/issues/2375 for more details.