Cleans up the oximeter `Producer` trait
The trait `oximeter::Producer` was previously a bit unwieldy and
difficult to use in some contexts, like Dropshot servers. Specifically,
the trait didn't have a few super traits which made it impossible to
_share_ safely between code. But in the case of HTTP servers, and
probably others, we want to generate metrics in some thread/task, and
consume them in another (the producer server). Because of this, the
`Collector` type may also have been unsound. It stores boxed `Producer`
trait objects, but that trait doesn't require `Send`, `Sync`, or
`'static`, which means that the objects can't be exposed in a Dropshot
context. I unsafely implemented `Sync` and `Send` for that type, without
without really understanding why that was required.

In any case, we now express those trait bounds on the `Producer` trait,
which allows us to export them via Dropshot safely, and enables the case
where we produce and consume metrics concurrently, possibly from
different execution contexts. Note that this would still require people
to implement interior mutability for their type implementing `Producer`,
but that's no different from any situation in which one wants to safely
share mutable data between execution contexts.