Refactors oximeter into separate crates by function (#297)
* Refactors oximeter into separate crates by function

The oximeter crate was unwieldy and too large, requiring any crate
implementing even simple functionality like a new metric to link against
the entire package. This splits the oximeter crate into a subcomponents
with narrower functionality.

- `oximeter` retains the core, such as metric/target macros, traits, and
  core metric types like the histogram.
- `oximeter-export` contains code for exporting data to the rest of the
  control plane -- Dropshot endpoint handlers and a batteries-included
  Dropshot server (for applications which don't otherwise run their own).
- `oximeter-collector` implements the actual `oximeter` collection
  server. This is the architectural component called "oximeter" in the
  control plane reference documentation such as RFD 61.
- `oximeter-db` implements a client to the ClickHouse telemetry
  database. The vanilla `Client` object implements read functionality,
  for querying the database, retrieving timeseries and their schema. The
  write functionality is isolated in the `DbWrite` trait -- this is
  intended to be used only by the `oximeter-collector` crate, but in
  theory could be used by any consuming code.

Aside from moving crates and cleaning up the dependencies, this commit
is a non-functional change. The goal is to clean up the oximeter
components, and to make it easier for other parts of the control plane
to depend on the specific functionality they need.

* Address PR review, rename oximeter-export -> oximeter-producer, for
clarity and matching RFD terminology.