WIP conversation starter for OPTE + Sled Agent
I have not tested this nor even tried to compile it. I just wrote some
code on the fly to serve as a straw man for how I think Sled Agent +
OPTE interaction might look for the next demo milestone. In this
commit I only cover the registration/unregistration of the OPTE Port,
which is what overlays the illumos VNIC. Once again, this is messy and
stream of conciousness. Just wanna get the juices flowing in terms of
integration.