Adds timeseries name to oximeter query implementation
The previous implementation incorrectly filtered timeseries on the basis
of their timeseries keys and field names/values only. This isn't quite
right, as many timeseries can share the same key if they happen to have
the same values for the fields. What is unique is the combination of
timeseries name and timeseries key. This adds the timeseries name to the
internal WHERE clause when selecting samples, and the other places
timeseries are expected to be unique. The returned samples are also
ordered by the name, in addition to the key and timestamp as before.