Add code for creating, splitting, combining a RackSecret
A new `RackSecret` type used to implement the "Trust Quorum" functionality of
the rack is provided in this commit. It uses Shamir secret sharing with Feldman
verification. Feldman verification was chosen for simplicity as it doesn't
require creating blinding shares and distributing those along with the actual
shares as with Pederson verification. We can always change this later.

The NIST P-256 curve was chosen somewhat arbitrarily, although it is a popular
curve and works with Shamir secret sharing without any extra mechanisms required
by some other curves.

A very simple versioned protocol will be created to use this functionality over
SPDM channels. The use of a P-256 curve, Feldman verification, and bincode for
serialization will serve as version 1. It's unlikely there will be a need to
change this in the near future. Upgrade can be supported by transitioning to a
later version once all members of the quorum support the latest version of the
code.